% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoElgp.R
\name{autoELGP}
\alias{autoELGP}
\title{AUTO-ELGP}
\usage{
autoELGP(x, y, t, p, min = NULL, max = NULL, s = NULL)
}
\arguments{
\item{x}{variable}

\item{y}{response variable}

\item{t}{variable exposure (t>>y)}

\item{p}{degree of polynomial}

\item{min}{minimum bandwidth}

\item{max}{maximum bandwidth}

\item{s}{increment for bandwidth searching}
}
\description{
AUTO-ELGP
}
\details{
autoElgp
}

