% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELGP.R
\name{ELGP}
\alias{ELGP}
\title{ELGP}
\usage{
ELGP(x, y, t, h, p)
}
\arguments{
\item{x}{variable}

\item{y}{response var}

\item{t}{variable Exposure (t>>y)}

\item{h}{bandwith}

\item{p}{degree of polynomial (0,1)}
}
\description{
ELGP
}

