% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makcv.R
\name{Makcv}
\alias{Makcv}
\title{MAKCV}
\usage{
Makcv(x, y, t, min = 1, max = 5, s = 1, p = 0)
}
\arguments{
\item{x}{variable}

\item{y}{response variable}

\item{t}{exposure variable (t>>y)}

\item{min}{minimum bandwidth}

\item{max}{maximum bandwidth}

\item{s}{step bandwidth range}

\item{p}{degree of polynomial (0,1)}
}
\description{
MAKCV
}
\details{
MAKCV: find best bandwith via cross validation
}

