% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.elgp.R
\name{plotElgp}
\alias{plotElgp}
\title{Plot ELGP}
\usage{
plotElgp(x, y, t, elgpres, addlegend = F, add = F, ...)
}
\arguments{
\item{x}{variable}

\item{y}{respons variale}

\item{t}{exposyre variable}

\item{elgpres}{elgp result}

\item{addlegend}{boolean for add legend on graphics}

\item{add}{boolean for adding other plot}

\item{...}{other graphical parameter}
}
\description{
Plot ELGP
}
\details{
Plotting elgp result
}

