% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGUI.R
\name{plotGUI}
\alias{plotGUI}
\title{PLOT GUI}
\usage{
plotGUI(x, y, t)
}
\arguments{
\item{x}{variable}

\item{y}{response variable}

\item{t}{variable exposure}
}
\description{
PLOT GUI
}
\details{
GUI simulation for ELGP
}

